refactor benchmarks; use standard set similarity search datasets; add HNSW benchmark
