+++ benchmark/index/requirements.txt
@@ -1,0 +1,4 @@
+pyfarmhash
+nmslib
+matplotlib
+SetSimilaritySearch
