+++ benchmark/lshensemble/requirements.txt
@@ -1,0 +1,5 @@
+SetSimilaritySearch
+matplotlib
+pandas
+farmhash
+scipy
