+++ benchmark/sketch/similarity_benchmark.py
